# src/main/java/com/northbay/rag_chat_storage/controller/ChatMessageController.java
package com.northbay.rag_chat_storage.controller;

import java.util.UUID;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import com.northbay.rag_chat_storage.annotations.RateLimited;
import com.northbay.rag_chat_storage.dto.ChatMessageRequest;
import com.northbay.rag_chat_storage.models.ChatMessage;
import com.northbay.rag_chat_storage.models.ChatSession;
import com.northbay.rag_chat_storage.service.ChatMessageService;
import com.northbay.rag_chat_storage.service.ChatSessionService;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/api/v1/sessions/{sessionId}/messages")
@RequiredArgsConstructor
public class ChatMessageController {

    @Autowired
    private  ChatMessageService messageService;

    @Autowired
    private  ChatSessionService sessionService;

    private static final String LIMITER_NAME = "chatMessageLimiter";

    /**
     * Add a new message to a chat session.
     */
    @RateLimited(name = LIMITER_NAME)
    @PostMapping
    public ResponseEntity<ChatMessage> addMessage(
            @PathVariable UUID sessionId,
            @Valid @RequestBody ChatMessageRequest request) {

        ChatSession session = sessionService.getSession(sessionId);
        ChatMessage savedMessage = messageService.addMessage(
                session,
                request.getSender(),
                request.getContent(),
                request.getContext());

        return ResponseEntity.status(HttpStatus.CREATED).body(savedMessage);
    }

    /**
     * Fetch paginated chat messages for a session.
     */
    @RateLimited(name = LIMITER_NAME)
    @GetMapping
    public ResponseEntity<Page<ChatMessage>> getMessages(
            @PathVariable UUID sessionId,
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "10") int size) {

        ChatSession session = sessionService.getSession(sessionId);
        Page<ChatMessage> messages = messageService.getMessages(session, PageRequest.of(page, size));

        return ResponseEntity.ok(messages);
    }
    
    @RateLimited(name = LIMITER_NAME)
    @DeleteMapping("/{messageId}")
    public ResponseEntity<Void> deleteMessage(
            @PathVariable UUID sessionId,
            @PathVariable UUID messageId) {

        ChatSession session = sessionService.getSession(sessionId);
        messageService.deleteMessage(session, messageId);

        return ResponseEntity.noContent().build();
    }
}